\name{dieu-package}
\alias{dieu-package}
\alias{dieu}
\docType{package}
\title{
\packageTitle{dieu}
}
\description{
\packageDescription{dieu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dieu}
\packageIndices{dieu}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{dieu}

Maintainer: \packageMaintainer{dieu}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
