% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dieu.R
\docType{package}
\name{dieu}
\alias{dieu}
\alias{dieu-package}
\title{dieu: euclidan and dijkstra algorithm}
\description{
conduct euclidean and dikstra algorithm for given input
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\author{
Maintainer: Josh Hyungyum Kim gyommy15@gmail.com
}
