% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Applying Dijkstra algorithm to a given graph and initial node}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame that contains 3 vectors named v1, v2, w}

\item{init_node}{Numeric representation of initial node}
}
\value{
Solution of the Dijkstra algorithm from the given initial node
}
\description{
Applying Dijkstra algorithm to a given graph and initial node
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
