% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Greatest Common Dvisor(GCD) for two input variables}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number(integer)}

\item{y}{A number(integer)}
}
\value{
GCD(Greatest Common Divisor) of x and y
}
\description{
Calculate Greatest Common Dvisor(GCD) for two input variables
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
