% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example data set for Dijkstra algorithm in the package}
\format{A data frame with 3 columns with 18 variables:
\describe{
  \item{v1}{index of starting node}
  \item{v2}{index of destination node}
  \item{w}{weight of the edge from column1 to column2 component}
  ...
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the information of origin node, destination node and weight
}
\keyword{datasets}
