\name{dieu-package}
\alias{dieu-package}
\alias{dieu}
\docType{package}
\title{
\packageTitle{dieu}
}
\description{
\packageDescription{dieu}
}
\details{ conduct euclidean and dikstra algorithm for given input

}
\author{
\packageAuthor{dieu}

Maintainer: \packageMaintainer{dieu}
}
\references{
Euclidean algorithm: \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
 
Dijkstra algorithm: \url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{ package }
\seealso{
}
\examples{
}
